improvement: add dry-run option.

+++ exifsort.sh
@@ -41,6 +41,10 @@ FILETYPES=("*.jpg" "*.jpeg" "*.png" "*.tif" "*.tiff" "*.gif" "*.xcf" "*.mp4" "*.
 #
 DIR_BLACKLIST=('*lost*' '*noexif*' '*duplicates*' '*slideshows*' '*raw*')
 #
+# The following will make no changes unless DRY_RUN is set to false.
+#
+DRY_RUN="true"
+#
 # Optional: Prefix of new top-level directory to kjmove sorted photos to.
 # if you use MOVETO, it MUST have a trailing slash! Can be a relative pathspec, but an
 # absolute pathspec is recommended.
@@ -124,7 +128,9 @@ if [[ "$1" == "doAction" && "$2" != "" ]]; then
         else
             echo " Moving to ./noexif/"
             echo mkdir -p "${MOVETO}noexif" && echo mv -n "$2" "${MOVETO}noexif/"
-            mkdir -p "${MOVETO}noexif" && mv -n "$2" "${MOVETO}noexif/"
+            if [[ "$DRY_RUN" == "false" ]]; then
+                mkdir -p "${MOVETO}noexif" && mv -n "$2" "${MOVETO}noexif/"
+            fi
             exit
         fi
     else
@@ -186,8 +192,10 @@ if [[ "$1" == "doAction" && "$2" != "" ]]; then
     # Fix permissions
     chmod 644 "$2"
 
-    echo mkdir -p "${MOVETO}${DIRNAME}" && echo mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
-    mkdir -p "${MOVETO}${DIRNAME}" && mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
+    echo DEBUG: mkdir -p "${MOVETO}${DIRNAME}" && echo mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
+    if [[ "$DRY_RUN" == "false" ]]; then
+        mkdir -p "${MOVETO}${DIRNAME}" && mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
+    fi
     echo -e "INFO: done.\n"
     exit
 fi;
